+++ src/packets.rs
@@ -227,6 +227,43 @@ impl Idle {
     }
 }
 
+/// Instruct all decoders to stop, either immediately or by simply
+/// removing motor power ("float"). The specification allows a direction
+/// field, but that is not implemented here because in situations where
+/// a BroadcastStop is being sent it is unlikely that direction settings
+/// will be important. (whereas a regular stop might wish to retain
+/// headlight states)
+pub struct BroadcastStop {
+    float: bool,
+}
+
+impl BroadcastStop {
+    /// Bring all locomotives to an immediate stop
+    pub fn immediate() -> Self {
+        Self { float: false }
+    }
+
+    /// Bring all locomotives to a gently/floating stop
+    pub fn float() -> Self {
+        Self { float: true }
+    }
+
+    /// Serialise the packed into the provided buffer
+    pub fn serialise(&self, buf: &mut SerialiseBuffer) -> Result<usize> {
+        let instr = if self.float { 0b0101_0000 } else { 0b0100_0000 };
+        buf[0..16].copy_from_bitslice([0xff, 0xfe].view_bits::<Msb0>()); // preamble
+        buf.set(15, false); // start bit
+        buf[16..24].copy_from_bitslice([0x00].view_bits::<Msb0>());
+        buf.set(24, false); // data start bit
+        buf[25..33].copy_from_bitslice([instr].view_bits::<Msb0>());
+        buf.set(33, false); // ecc start bit
+        buf[34..42].copy_from_bitslice([instr].view_bits::<Msb0>());
+        buf.set(42, true); // stop bit
+
+        Ok(43)
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
@@ -357,6 +394,34 @@ mod test {
             0b0_0000_000, // start + instr[..7]
             0b0_0_111111, // instr[7] + start + ecc[..6]
             0b11_1_00000, // ecc[6..] + stop + 5 zeroes
+        ];
+
+        let mut expected = SerialiseBuffer::default();
+        expected[..43]
+            .copy_from_bitslice(&expected_arr.view_bits::<Msb0>()[..43]);
+        println!("got:");
+        display_serialise_buffer(&buf);
+        println!("expected:");
+        display_serialise_buffer(&expected);
+        assert_eq!(len, 43);
+        assert_eq!(buf[..len], expected[..43]);
+        Ok(())
+    }
+
+    #[test]
+    fn serialise_broadcast_stop_packet() -> Result<()> {
+        let pkt = BroadcastStop::float();
+        let mut buf = SerialiseBuffer::default();
+        let len = pkt.serialise(&mut buf)?;
+
+        #[allow(clippy::unusual_byte_groupings)]
+        let expected_arr = [
+            0xff_u8,      // preamble
+            0b1111_1110,  // preamble + start
+            0x00,         // address
+            0b0_0101_000, // start + instr[..7]
+            0b0_0_010100, // instr[7] + start + ecc[..6]
+            0b00_1_00000, // ecc[6..] + stop + 5 zeroes
         ];
 
         let mut expected = SerialiseBuffer::default();